+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The document follows the conventions described in [“Keep a CHANGELOG”](http:
 ## [2.6.2] - UNRELEASED
 
 ### Added
+- added `width()` and `height()` methods to `SVG.FX`
 
 ### Changed
 - updated Jasmine from 2.5.2 to 2.6.0